rearranged functions added skeleton of new functions (need hardware to fully setup)
